\name{FieldSpectroscopyDP-package}
\alias{FieldSpectroscopyDP-package}
\alias{FieldSpectroscopyDP}
\docType{package}
\title{R package for processing fieldspectroscopy data}
}
\description{R package for Characterization and Calibration of spectrometers}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FieldSpectroscopyDP}
\packageIndices{FieldSpectroscopyDP}
}
\author{Tommaso Julitta, Mirco Migliavacca, Thomas Wutzler}

Maintainer: \packageMaintainer{FieldSpectroscopyDP}
}

\keyword{ package }


