\name{log10_ceiling}
\alias{log10_ceiling}
\title{log10 ceiling}
\description{Round to the next order fo magnitude}
\usage{log10_ceiling(x)}
\arguments{
  \item{x}{numeric value or vector: number to be rounded}
}

\value{numeric value or vector. rounded to the next order of magnitude }

\author{Tommaso Julitta, Mirco Migliavacca, Thomas Wutzler}





